backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.8"
fastapi = "*"
sqlalchemy = "*"
asyncpg = "*"
greenlet = "*"
psycopg2 = "*"
uvicorn = "*"
openai = "^1.61.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
